% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code}
\alias{code}
\title{Insert a code block}
\usage{
code(code)
}
\arguments{
\item{code}{The code to embed in an HTML page.}
}
\description{
Generates a code block to embed in an HTML page.
}
\examples{
code('print("Hello, world.")')

}
