% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image}
\alias{image}
\title{Create a HTML reference to an existing image}
\usage{
image(image_name, image_dims = c(800, 800))
}
\arguments{
\item{image_name}{The filename of the image.}

\item{image_dims}{The x, y dimensions (in pixels) at which to display the image.}
}
\description{
Generates HTML markup to display an image.
}
\examples{
# image('image1.png', c(500,500))

}
