% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Insert a link}
\usage{
link(name, path)
}
\arguments{
\item{name}{The name of the link, as it is displayed in the HTML page.}

\item{path}{The path to which the link refers.}
}
\description{
Insert a link to a page, image or object.
}
\examples{
link('Click here to return to page 1.', path = 'page_1.html')

}
