% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math}
\alias{math}
\title{Write math equations using MathJax}
\usage{
math(equation)
}
\arguments{
\item{text}{The mathjax to embed in an HTML page.}
}
\description{
Write mathematical equation to embed in an HTML page. Note: special characters need to be escaped.
}
\examples{
math('$$ {J(\\\\theta) =\\\\frac{1}{2m} [\\\\sum^m_{i=1}(h_\\\\theta(x^{(i)}) - y^{(i)})2 + \\\\lambda\\\\sum^n_{j=1}\\\\theta^2_j} $$')

}
