% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page}
\alias{page}
\title{Create an HTML page}
\usage{
page(...)
}
\arguments{
\item{...}{Elements to include on the page, such as a table(), image() and text().}

\item{filename}{The filename to which the generated HTML page will be written.}

\item{css}{A CSS style sheet to use for the formatting of the HTML page.}
}
\description{
Creates an HTML page based on several elements.
}
\examples{
page(
# title('Table'),
# table(df1),
# text('This is a table with some stuff'),
# code('code <- chunk'),
# image('image1.png', image_dims = c(500, 500)),
# text('Some more text.'),
# link('Go back to main page..','Table_of_contents.html'),
# )

}
