% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus2.R
\name{+.page}
\alias{+.page}
\title{An operator to add a page to an existing page}
\usage{
\method{+}{page}(...)
}
\value{
A 'page' object.
}
\description{
Adds a page to the end of a page
}
\examples{
page1 <- page(
title('This is the title'),
tab(matrix(rnorm(100), nrow = 10, ncol = 10))
)

page1 <- page1 + page1
page1
}
