% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{+.pageObject}
\alias{+.pageObject}
\title{An operator to add elements to an existing page}
\usage{
\method{+}{pageObject}(...)
}
\value{
A 'page' object.
}
\description{
Adds elements to the end of a page
}
\examples{
page1 <- page(
title('This is the title'),
tab(matrix(rnorm(100), nrow = 10, ncol = 10))
)

page1 <- page1 + tab(matrix(rnorm(100), nrow = 10, ncol = 10))
page1

page1 <- page1 + text('Just wanted to add some text')
page1

}
