% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render HTML pages}
\usage{
render(
  ...,
  filename = NULL,
  css = list.files(system.file("extdata", package = "htmlbuddy"), full.names = TRUE)
)
}
\arguments{
\item{...}{Page(s) to be rendered. Objects of class 'page', separated by commas.}

\item{filename}{The filename(s) to which the generated HTML page will be written. Pass as character vector.}

\item{css}{A CSS style sheet to use for the formatting of the HTML page(s).}
}
\description{
Renders one or multiple pages to HTML files
}
\examples{
# render(
#    page1,
#    page2,
#    page3,
#    filename = paste0('page_', 1:3)
#    )

}
