% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\title{Add a side bar to a page}
\usage{
sidebar(..., side = "left", small = TRUE)
}
\arguments{
\item{...}{The content to embed inside the bar}

\item{side}{Controls on which side of the screen the sidebar should be inserted.}

\item{small}{Displays the font size smaller than normal text.}
}
\description{
Adds a side bar to a page, on the left or right side.
}
\examples{
# page + sidebar(text('Content')) # adds side bar on the left side of the page
#'page + sidebar(text('Content'), side = 'right') # adds side bar on the right side of the page

}
