% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Print a data.frame or matrix as HTML table}
\usage{
tab(object, escape = TRUE)
}
\arguments{
\item{object}{The table to be printed.}

\item{escape}{Argument passed to kable() to escape special characters.}
}
\description{
Outputs a data.frame or matrix as HTML table.
}
\examples{
m <- matrix(rnorm(50), nrow = 10, ncol = 5)
tab(m)

}
