% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Print a data.frame or matrix as HTML table}
\usage{
table(object, escape = TRUE)
}
\arguments{
\item{object}{The table to be printed.}

\item{escape}{Argument passed to kable() to escape special characters.}
}
\description{
Outputs a data.frame or matrix as HTML table.
}
\examples{
m <- matrix(rnorm(50), nrow = 10, ncol = 5)
table(m)

}
