% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text}
\alias{text}
\title{Write text}
\usage{
text(text)
}
\arguments{
\item{text}{The text to embed in an HTML page.}
}
\description{
Write text to embed in an HTML page.
}
\examples{
text('Hello, world.')

}
