% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{Create a table of contents}
\usage{
toc(header, page_titles, page_links)
}
\arguments{
\item{header}{The name of the header.}

\item{page_titles}{The names of the pages.}

\item{page_links}{The filenames to which is linked.}
}
\description{
Makes a table of contents with one header and several sub-items linked underneath.
}
\examples{
toc('Header 1', c('page 1','page 2'), c('page_1.html','page_2.html'))

}
