% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{Batch execute a function}
\usage{
batch(list, expr)
}
\arguments{
\item{list}{The list object, containing objects (e.g., data.frames, character vectors, etc.)}

\item{expr}{The function to be applied, e.g., tab, image, or text.}
}
\description{
Perform a given function on a list
}
\examples{
table_list <- list(head(iris), tail(iris))
batch(table_list, tab)

# this outputs the table HTML for both tables in the list

}
