% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title}
\alias{title}
\title{Display a title}
\usage{
title(text)
}
\arguments{
\item{text}{The title to embed in an HTML page.}
}
\description{
Display text formatted as a title.
}
\examples{
title('This is the title.')

}
